move common/complete to the only place it will ever be used
